+++ mhrb/js/model/calculate/index.js
@@ -256,6 +256,8 @@ function calculateBuildPerformance(db, build, calcState) {
 
     assert(m.reloadSpeedAdd !== undefined);
 
+    assert(m.sharpnessLevelReduction !== undefined);
+
     //
     // STAGE 2: Find Sharpness Modifiers
     //
@@ -286,7 +288,7 @@ function calculateBuildPerformance(db, build, calcState) {
         const effectiveHandicraftLevel = (barIsFull) ? 5 : s.handicraftLevel;
 
         // Now, we apply this effective handicraft level.
-        const sharpnessValues = getSharpnessValues(b.maxSharpness, effectiveHandicraftLevel);
+        const sharpnessValues = getSharpnessValues(b.maxSharpness, effectiveHandicraftLevel, m.sharpnessLevelReduction);
         realSharpnessBar           = sharpnessValues.realSharpnessBar;
         maxSharpnessBar            = b.maxSharpness;
         rawSharpnessModifier       = sharpnessValues.rawSharpnessModifier;
@@ -375,6 +377,14 @@ function calculateBuildPerformance(db, build, calcState) {
     //
     // STAGE 6: Apply Master's Touch and Razor Sharp to the sharpness bar
     //
+
+    // Innate razor sharp
+    switch (weaponRO.category) {
+        case "dualblades":   hitsMultiplier *= 3;   break; // 2/3 razor sharp, i.e. 1 of 3 hits reduces sharpness
+        case "insectglaive": hitsMultiplier *= 1.5; break; // 1/3 razor sharp, i.e. 2 of 3 hits reduces sharpness
+        default:
+            // Do nothing
+    }
 
     if (s.razorSharpLevel > 0) {
         const chanceOfNoSharpnessLoss = (()=>{
@@ -506,8 +516,18 @@ function calculateBuildPerformance(db, build, calcState) {
 /****************************************************************************************/
 
 
-function getSharpnessValues(maxSharpness, handicraftLevel) {
+function getSharpnessValues(maxSharpness, handicraftLevel, sharpnessLevelReduction) {
     const realSharpnessBar = applyHandicraft(maxSharpness, handicraftLevel);
+
+    let r = sharpnessLevelReduction;
+    for (let i = realSharpnessBar.length - 1; i > 0; --i) {
+        if (r <= 0) break;
+        if (realSharpnessBar[i] > 0) {
+            realSharpnessBar[i] = 0;
+            --r;
+        }
+    }
+
     const highestSharpnessLevel = getHighestSharpnessIndex(realSharpnessBar);
     return {
         realSharpnessBar:           realSharpnessBar,